feat: #748 토론 상세 화면 기능 추가 및 UIState 리팩터링

+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/response/discussion/DiscussionResponse.kt
@@ -12,9 +12,9 @@ data class DiscussionResponse(
     val discussionId: Long,
     val discussionOpinion: String,
     val discussionTitle: String,
-    val viewCount: Int,
-    val likeCount: Int,
-    val commentCount: Int,
+    val viewCount: Long,
+    val likeCount: Long,
+    val commentCount: Long,
     val isLikedByMe: Boolean,
 )
